move vec code into its own source file
